Adapt to Android Studio, change upload method to openConnection() and reqeue failed uploads
