+++ .gitignore
@@ -1,0 +1,9 @@
+
+/.idea/
+*.iml
+
+build/
+.gradle/
+
+/local.properties
+
