+++ app/src/main/AndroidManifest.xml
@@ -20,10 +20,6 @@
         <service android:name="net.zekjur.davsync.UploadService" />
 
         <activity
-            android:name="net.zekjur.davsync.MainActivity"
-            android:label="@string/app_name" >
-        </activity>
-        <activity
             android:name="net.zekjur.davsync.ShareActivity"
             android:excludeFromRecents="true"
             android:label="@string/title_activity_share"
