+++ app/src/main/java/net/zekjur/davsync/SettingsActivity.java
