+++ app/src/main/java/net/zekjur/davsync/ShareActivity.java
