+++ app/src/main/res/values/strings_activity_settings.xml
