+++ app/src/main/res/xml/pref_autosync.xml
